feat: add custom new player page #165
